refactor: PR 반영
