+++ src/main/kotlin/blackjack/domain/Number.kt
@@ -14,6 +14,4 @@ enum class Number(val symbol: String, val value: Int, val orValue: Int = 0) {
     JACK("J", 10),
     QUEEN("Q", 10),
     KING("K", 10);
-
-    fun isAce() = this == ACE
 }
