+++ src/test/kotlin/blackjack/domain/CardsTest.kt
@@ -57,4 +57,28 @@ class CardsTest : StringSpec({
             cards.point() shouldBe point
         }
     }
+
+    "Ace 는 블랙잭 숫자보다 합산이 작을 경우엔 포인트가 11로 계산된다." {
+        val cards = Cards(
+            listOf(
+                Card(Suit.SPADE, Number.TWO),
+                Card(Suit.HEART, Number.TWO),
+                Card(Suit.SPADE, Number.ACE)
+            )
+        )
+
+        cards.point() shouldBe 15
+    }
+
+    "Ace 는 블랙잭 숫자보다 합산이 클 경우엔 포인트가 1로 계산된다." {
+        val cards = Cards(
+            listOf(
+                Card(Suit.HEART, Number.NINE),
+                Card(Suit.DIAMOND, Number.NINE),
+                Card(Suit.SPADE, Number.ACE)
+            )
+        )
+
+        cards.point() shouldBe 19
+    }
 })
